backendRAG/requirements.txt
fastapi
uvicorn
python-dotenv
cohere
qdrant-client
pydantic
Pillow
PyMuPDF
google-genai
python-multipart
aiofiles
numpy

